+++ src/app/api/openai/tts/route.ts
@@ -1,9 +1,11 @@
 import { OpenAITTSPayload } from '@lobehub/tts';
 import { createOpenaiAudioSpeech } from '@lobehub/tts/server';
 
+import { getPreferredRegion } from '../../config';
 import { createBizOpenAI } from '../createBizOpenAI';
 
 export const runtime = 'edge';
+export const preferredRegion = getPreferredRegion();
 
 export const POST = async (req: Request) => {
   const payload = (await req.json()) as OpenAITTSPayload;
